backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic-settings==2.10.1
python-dateutil==2.9.0.post0
